Moved GCS and drone data into respective files
